refactor: split large modules into sub-files (Phase 1, Issue #575)
- Create thinking-constants.ts to unify THINKING_TAIL_LINE_COUNT across modules
- Split prompt-detector.ts (973 lines) into types.ts and prompt-detect-multiple-choice.ts
- Split db-migrations.ts (1300 lines) into migrations/ directory with runner + 5 version groups
- Split response-poller.ts (958 lines) into response-poller-core.ts and response-checker.ts
- Move normalizePromptForDedup to prompt-dedup.ts
- All existing import paths preserved via barrel re-exports
- All 286 test files pass (5540 tests), lint clean, build success
